requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysqlclient==2.1.1
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
